Say what a keyed table still owes the store contract, and share the fake provider
The appendMessages doc said a table keyed by message id "gets this for
free", which reads as "insert is enough" — and a plain insert of the
amended message hits the very key violation the contract exists to
prevent. What the key buys is the lookup; the write is still an
insert-or-update. The comment now says so.

The controller test had grown its own copy of the scripted provider
that Agent.test.ts already had. It now lives in providers/fakeProvider.ts,
in source rather than a test file for the reason store/stubAgentRun.ts
gives, and both test files import it.
